contract-review-tool/requirements.txt
flask
anthropic
pypdf
python-docx
werkzeug
python-dotenv
